# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Motor Clínico - API para clínica médica integrativa
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pacientes:
    get:
      operationId: listarPacientes
      tags: [pacientes]
      summary: Listar pacientes
      parameters:
        - name: unidadeId
          in: query
          schema:
            type: integer
        - name: busca
          in: query
          schema:
            type: string
        - name: pagina
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Lista de pacientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Paciente"
    post:
      operationId: criarPaciente
      tags: [pacientes]
      summary: Criar paciente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarPacienteBody"
      responses:
        "201":
          description: Paciente criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paciente"
  /pacientes/{id}:
    get:
      operationId: obterPaciente
      tags: [pacientes]
      summary: Obter paciente por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Paciente encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paciente"
    put:
      operationId: atualizarPaciente
      tags: [pacientes]
      summary: Atualizar paciente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarPacienteBody"
      responses:
        "200":
          description: Paciente atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Paciente"

  # Anamnese
...
  /motor-clinico/sugestoes:
    get:
      operationId: listarSugestoes
      tags: [motor-clinico]
      summary: Listar sugestões pendentes de validação
      parameters:
        - name: pacienteId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pendente, validado, rejeitado]
        - name: tipo
          in: query
          schema:
            type: string
            enum: [exame, formula, injetavel, implante, protocolo]
      responses:
        "200":
          description: Lista de sugestões
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SugestaoClinica"
  /motor-clinico/sugestoes/{id}/validar:
    post:
      operationId: validarSugestao
      tags: [motor-clinico, validacao]
      summary: Validar ou rejeitar sugestão do motor clínico
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidarSugestaoBody"
      responses:
        "200":
          description: Sugestão validada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SugestaoClinica"

  # Itens Terapêuticos - Toggles
  /motor-clinico/itens-terapeuticos:
    get:
      operationId: listarItensTerapeuticos
      tags: [motor-clinico]
      summary: Listar itens terapêuticos com toggles de disponibilidade
      parameters:
        - name: categoria
          in: query
          schema:
            type: string
            enum: [formula, injetavel_im, injetavel_ev, implante, exame, protocolo]
        - name: disponivel
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Lista de itens terapêuticos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ItemTerapeutico"
    post:
      operationId: criarItemTerapeutico
      tags: [motor-clinico]
      summary: Criar item terapêutico
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarItemTerapeuticoBody"
      responses:
        "201":
          description: Item criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemTerapeutico"
  /motor-clinico/itens-terapeuticos/{id}/toggle:
    patch:
      operationId: toggleItemTerapeutico
      tags: [motor-clinico]
      summary: Ativar/desativar item terapêutico
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleItemBody"
      responses:
        "200":
          description: Toggle atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemTerapeutico"

  # Protocolos
...
  /followup:
    get:
      operationId: listarFollowups
      tags: [followup]
      summary: Listar follow-ups agendados
      parameters:
        - name: pacienteId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [agendado, realizado, cancelado, atrasado]
        - name: unidadeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Lista de follow-ups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Followup"
    post:
      operationId: criarFollowup
      tags: [followup]
      summary: Criar follow-up
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarFollowupBody"
      responses:
        "201":
          description: Follow-up criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Followup"
  /followup/{id}/concluir:
    post:
      operationId: concluirFollowup
      tags: [followup]
      summary: Marcar follow-up como concluído
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConcluirFollowupBody"
      responses:
        "200":
          description: Follow-up concluído
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Followup"

  # Financeiro
  /financeiro/pagamentos:
    get:
      operationId: listarPagamentos
      tags: [financeiro]
      summary: Listar pagamentos
      parameters:
        - name: pacienteId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pendente, pago, cancelado, estornado]
        - name: unidadeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Lista de pagamentos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pagamento"
    post:
      operationId: criarPagamento
      tags: [financeiro]
      summary: Criar registro de pagamento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CriarPagamentoBody"
      responses:
        "201":
          description: Pagamento criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pagamento"
  /financeiro/pagamentos/{id}/confirmar:
    post:
      operationId: confirmarPagamento
      tags: [financeiro]
      summary: Confirmar pagamento
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pagamento confirmado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pagamento"

  # Unidades
...
components:
  schemas:
...
    Paciente:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        cpf:
          type: string
        dataNascimento:
          type: string
          format: date
        telefone:
          type: string
        email:
          type: string
        unidadeId:
          type: integer
        statusAtivo:
          type: boolean
        criadoEm:
          type: string
          format: date-time
        atualizadoEm:
          type: string
          format: date-time
      required:
        - id
        - nome
        - telefone
        - unidadeId
        - statusAtivo
        - criadoEm
        - atualizadoEm
    CriarPacienteBody:
      type: object
      properties:
        nome:
          type: string
        cpf:
          type: string
        dataNascimento:
          type: string
          format: date
        telefone:
          type: string
        email:
          type: string
        unidadeId:
          type: integer
      required:
        - nome
        - telefone
        - unidadeId
...
    SugestaoClinica:
      type: object
      properties:
        id:
          type: integer
        anamneseId:
          type: integer
        pacienteId:
          type: integer
        pacienteNome:
          type: string
        tipo:
          type: string
          enum: [exame, formula, injetavel_im, injetavel_ev, implante, protocolo]
        itemTerapeuticoId:
          type: integer
        itemNome:
          type: string
        itemDescricao:
          type: string
        justificativa:
          type: string
        prioridade:
          type: string
          enum: [baixa, media, alta, urgente]
        status:
          type: string
          enum: [pendente, validado, rejeitado, em_execucao, concluido]
        validadoPorId:
          type: integer
        validadoPorNome:
          type: string
        validadoEm:
          type: string
          format: date-time
        observacaoValidacao:
          type: string
        criadoEm:
          type: string
          format: date-time
      required:
        - id
        - anamneseId
        - pacienteId
        - tipo
        - itemNome
        - prioridade
        - status
        - criadoEm
    ValidarSugestaoBody:
      type: object
      properties:
        acao:
          type: string
          enum: [validar, rejeitar]
        observacao:
          type: string
        validadoPorId:
          type: integer
      required:
        - acao
        - validadoPorId
    ItemTerapeutico:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        descricao:
          type: string
        categoria:
          type: string
          enum: [formula, injetavel_im, injetavel_ev, implante, exame, protocolo]
        subCategoria:
          type: string
        disponivel:
          type: boolean
        preco:
          type: number
          format: float
        tags:
          type: array
          items:
            type: string
        criadoEm:
          type: string
          format: date-time
      required:
        - id
        - nome
        - categoria
        - disponivel
        - criadoEm
    CriarItemTerapeuticoBody:
      type: object
      properties:
        nome:
          type: string
        descricao:
          type: string
        categoria:
          type: string
          enum: [formula, injetavel_im, injetavel_ev, implante, exame, protocolo]
        subCategoria:
          type: string
        disponivel:
          type: boolean
        preco:
          type: number
        tags:
          type: array
          items:
            type: string
      required:
        - nome
        - categoria
    ToggleItemBody:
      type: object
      properties:
        disponivel:
          type: boolean
      required:
        - disponivel
...
    Followup:
      type: object
      properties:
        id:
          type: integer
        pacienteId:
          type: integer
        pacienteNome:
          type: string
        tipo:
          type: string
          enum: [consulta, exame, procedimento, ligacao, mensagem]
        status:
          type: string
          enum: [agendado, realizado, cancelado, atrasado]
        dataAgendada:
          type: string
          format: date-time
        dataRealizada:
          type: string
          format: date-time
        observacoes:
          type: string
        recorrencia:
          type: string
          enum: [nenhuma, diaria, semanal, quinzenal, mensal, trimestral]
        responsavelId:
          type: integer
        responsavelNome:
          type: string
        unidadeId:
          type: integer
        criadoEm:
          type: string
          format: date-time
      required:
        - id
        - pacienteId
        - pacienteNome
        - tipo
        - status
        - dataAgendada
        - unidadeId
        - criadoEm
    CriarFollowupBody:
      type: object
      properties:
        pacienteId:
          type: integer
        tipo:
          type: string
          enum: [consulta, exame, procedimento, ligacao, mensagem]
        dataAgendada:
          type: string
          format: date-time
        observacoes:
          type: string
        recorrencia:
          type: string
          enum: [nenhuma, diaria, semanal, quinzenal, mensal, trimestral]
        responsavelId:
          type: integer
        unidadeId:
          type: integer
      required:
        - pacienteId
        - tipo
        - dataAgendada
        - unidadeId
    ConcluirFollowupBody:
      type: object
      properties:
        observacoes:
          type: string
        responsavelId:
          type: integer
      required:
        - responsavelId
    Pagamento:
      type: object
      properties:
        id:
          type: integer
        pacienteId:
          type: integer
        pacienteNome:
          type: string
        valor:
          type: number
          format: float
        status:
          type: string
          enum: [pendente, pago, cancelado, estornado]
        formaPagamento:
          type: string
          enum: [dinheiro, cartao_credito, cartao_debito, pix, boleto, plano_saude]
        descricao:
          type: string
        unidadeId:
          type: integer
        criadoEm:
          type: string
          format: date-time
        paguEm:
          type: string
          format: date-time
      required:
        - id
        - pacienteId
        - pacienteNome
        - valor
        - status
        - unidadeId
        - criadoEm
    CriarPagamentoBody:
      type: object
      properties:
        pacienteId:
          type: integer
        valor:
          type: number
        formaPagamento:
          type: string
          enum: [dinheiro, cartao_credito, cartao_debito, pix, boleto, plano_saude]
        descricao:
          type: string
        unidadeId:
          type: integer
      required:
        - pacienteId
        - valor
        - formaPagamento
        - unidadeId
...