# lib/api-zod/src/generated/types/concluirFollowupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Motor Clínico - API para clínica médica integrativa
 * OpenAPI spec version: 0.1.0
 */

export interface ConcluirFollowupBody {
  observacoes?: string;
  responsavelId: number;
}